Modificación de la firma de la clase InventarioDePrioridad
Se cambia el nombre de la clase de Inventario_de_prioridad a InventarioDePrioridad para respetar el formato CamelCase de las otras clases.

Se agrega el método vacío.

Se modifican algunas postcondiciones que no indicaban bien que hace el método.

Se crea una constante ARMA_NO_VALIDA la cual representa un arma no válida, util para considerar casos border como por ejemplo que el inventario esté vacío y no haya arma que devolver.

+++ src/Floyd.cpp
@@ -20,10 +20,10 @@ std::vector<size_t> Floyd::obtener_camino(size_t origen, size_t destino) {
     size_t actual = origen;
 
     while (!terminado) {
-        actual = matriz_caminos.elemento(actual, destino);
+        actual = size_t(matriz_caminos.elemento(actual, destino));
         camino.push_back(actual);
 
-        if (matriz_caminos.elemento(actual, destino) == destino) {
+        if (size_t(matriz_caminos.elemento(actual, destino)) == destino) {
             terminado = true;
             camino.insert(camino.begin(), origen);
             camino.push_back(destino);
